requirements.txt
playwright>=1.28.0,<2.0
PyYAML>=6.0,<7.0
